fix(mails): report a failed mark write instead of answering success
Review caught that the route discarded the boolean markRead/markSaved return.
The repositories catch their own errors and report the outcome, so a transient
DB failure answered success for a write that never landed — and for read=true
the badge had already been decremented for a mail that stayed unread. That is
the same answers-success-while-nothing-happened class this PR is fixing, so it
belongs here. The badge decrement now chains after a confirmed write.

The route mocks returned undefined rather than the real Promise<boolean>, which
is what let the ignored return value pass unnoticed; they now return true, and
new cases cover each write reporting false.

Also rejects a FilterCondition whose `value` is undefined — it bound NULL and
matched nothing rather than widening, but it is still a caller naming a
predicate it had no value for, and the docstring claimed more than the code
enforced.

Moves the mutation-filter tests into models.test.ts, which already owns base.ts's
pure helpers, rather than a where-filters.test.ts with no matching source file.
